added try catch in handler to catch throwable apps
