+++ zio-http/src/main/scala/zhttp/service/Handler.scala
@@ -23,28 +23,40 @@ private[zhttp] final case class Handler[R](
   override def channelRead0(ctx: Ctx, jReq: FullHttpRequest): Unit = {
     jReq.touch("server.Handler-channelRead0")
     implicit val iCtx: ChannelHandlerContext = ctx
-    unsafeRun(
-      jReq,
-      app,
-      new Request {
-        override def method: Method = Method.fromHttpMethod(jReq.method())
+    try
+      (
+        unsafeRun(
+          jReq,
+          app,
+          new Request {
+            override def method: Method = Method.fromHttpMethod(jReq.method())
 
-        override def url: URL = URL.fromString(jReq.uri()).getOrElse(null)
+            override def url: URL = URL.fromString(jReq.uri()).getOrElse(null)
 
-        override def headers: Headers = Headers.make(jReq.headers())
+            override def headers: Headers = Headers.make(jReq.headers())
 
-        override def unsafeEncode: HttpRequest = jReq
+            override def unsafeEncode: HttpRequest = jReq
 
-        override def remoteAddress: Option[InetAddress] = {
-          ctx.channel().remoteAddress() match {
-            case m: InetSocketAddress => Some(m.getAddress)
-            case _                    => None
-          }
-        }
+            override def remoteAddress: Option[InetAddress] = {
+              ctx.channel().remoteAddress() match {
+                case m: InetSocketAddress => Some(m.getAddress)
+                case _                    => None
+              }
+            }
 
-        override def data: HttpData = HttpData.fromByteBuf(jReq.content())
-      },
-    )
+            override def data: HttpData = HttpData.fromByteBuf(jReq.content())
+          },
+        ),
+      )
+    catch {
+      case throwable: Throwable =>
+        writeResponse(
+          Response
+            .fromHttpError(HttpError.InternalServerError(cause = Some(throwable)))
+            .withConnection(HeaderValues.close),
+          jReq,
+        ): Unit
+    }
   }
 
   /**
